[2단계 - 경로 조회 기능] 바론(이소민) 미션 제출합니다. (#151)
* docs: API 명세 수정 및 id 반환값 추가

* refactor: Line 반환값 필드명 변경

* refactor: 테스트 URI 변경

* refactor: LineService, Controller DTO 분리

* refactor: SectionService, Controller DTO 분리

* refactor: StationService, StationController DTO 분리

* docs: 삭제된 API 명세 제거

* docs: 최단 경로 및 요금 계산 기능 요구사항 작성

* feat: jgrapht를 사용하여 최단 경로 조회 도메인 생성

* feat: 출발역에서 도착역까지 최단 경로 조회 및 거리 반환

* feat: 경로 길이에 따른 요금 계산

* refactor: Section 조회 메서드명 변경

* feat: 최단 경로 조회 API 구현

* feat: 프로덕션, 테스트 데이터베이스 분리

* feat: 예외 문자열 response 포장

* refactor: 새 구간역 추가 후 추가된 구간 반환하도록 변경

* refactor: dto 패키지 controller, service 레이어로 분리

* refactor: 사용되지 않는 코드 제거

* refactor: 경로 조회 API URI 변경

- /routes -> /paths

* refactor: 경로 조회 API 출발역, 도착역 데이터 변경

- RequestBody -> RequestParam
- 역 이름 -> 역 id

* refactor: Sections 정렬된 구간 조회 메서드 분리

- Sections: getSections(), getSectionsByList() 삭제, getSortedSections() 추가
- SectionRepository: 전체 구간 조회 메서드 findALl() 추가

* refactor: 불필요한 SectionEntity 조회 메서드 삭제

* feat: Distance VO 생성

* feat: Fare VO 생성

* refactor: profile 설정

* refactor: nullable 한 wrapper -> primitive 로 변경

* refactor: Line final 추가

* refactor: response 변환 메서드 분리

* refactor: FareCalculator util 클래스로 변경

* refactor: Exception 에외 메세지 밖에서 주입

* refactor: Request -> DTO 통합

* style: 구글 자바 컨벤션 적용

+++ src/test/resources/schema.sql
@@ -4,26 +4,26 @@ DROP TABLE IF EXISTS LINE;
 
 create table STATION
 (
-    id bigint auto_increment not null,
+    id   bigint auto_increment not null,
     name varchar(255) not null unique,
     primary key (id)
 );
 
 create table LINE
 (
-    id                 bigint auto_increment not null,
-    name               varchar(255)          not null unique,
-    color              varchar(20)           not null,
+    id    bigint auto_increment not null,
+    name  varchar(255) not null unique,
+    color varchar(20)  not null,
     primary key (id)
 );
 
 create table SECTION
 (
-    id                 bigint auto_increment not null,
-    line_id            bigint                not null,
-    start_station_id   bigint                not null,
-    end_station_id     bigint                not null,
-    distance           int                   not null,
+    id               bigint auto_increment not null,
+    line_id          bigint not null,
+    start_station_id bigint not null,
+    end_station_id   bigint not null,
+    distance         double not null,
     primary key (id),
     foreign key (line_id) REFERENCES LINE (id),
     foreign key (start_station_id) REFERENCES STATION (id),
